+++ packages/core/types/server/launch-editor.d.ts
@@ -1,2 +1,3 @@
-declare function launchEditor(fileName: string, lineNumber: unknown, colNumber: unknown): void;
+import { Editor } from '../shared/constant';
+declare function launchEditor(fileName: string, lineNumber: unknown, colNumber: unknown, _editor?: Editor): void;
 export default launchEditor;